requirements.txt
certifi==2018.10.15
mkl-fft==1.0.6
mkl-random==1.0.1
numpy==1.15.4
olefile==0.46
opencv-python==3.4.4.19
Pillow==5.3.0
